dlvpass changes for same day
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@18266 842219ff-9341-0410-a943-573460bdd4de

+++ projects/ERPSWebApp/src/com/freshdirect/webapp/taglib/fdstore/DeliveryTimeSlotTag.java
@@ -227,7 +227,7 @@ public class DeliveryTimeSlotTag extends AbstractGetterTag<Result> {
 		List<FDTimeslotUtil> timeslotList = getFDTimeslotListForDateRange(restrictions, dateRanges,
 				result, timeslotAddress, user,event);
 		
-		if(!cart.isDlvPassPremiumAllowedTC())
+		if(cart.isDlvPassApplied() && !cart.isDlvPassPremiumAllowedTC())
 			TimeslotLogic.purgeSDSlots(timeslotList);
 		
 		showPremiumSlots =TimeslotLogic.hasPremiumSlots(timeslotList, baseRange.getStartDate(), DateUtil.addDays(baseRange.getEndDate(),-1));
